Правит типы предложений
