+++ src/types/offer.ts
@@ -1,31 +1,9 @@
-export type Location = {
-  latitude: number;
-  longitude: number;
-  zoom: number;
-};
-
-export type City = {
-  name: string;
-  location: Location;
-};
+import { Offers } from './offers';
 
 type Host = {
   name: string;
   avatarUrl: string;
   isPro: boolean;
-};
-
-export type Offers = {
-  id: string;
-  title: string;
-  type: string;
-  price: number;
-  city: City;
-  location: Location;
-  isFavorite: boolean;
-  isPremium: boolean;
-  rating: number;
-  previewImage: string;
 };
 
 type OfferInfo = {
