+++ src/types/offers.ts
@@ -1,14 +1,23 @@
-import { City } from './offer';
+export type Location = {
+  latitude: number;
+  longitude: number;
+  zoom: number;
+};
+
+export type City = {
+  name: string;
+  location: Location;
+};
 
 export type Offers = {
-  id: string;
-  title: string;
-  type: string;
-  price: number;
   city: City;
-  location: Location;
+  goods: string[];
+  id: string;
   isFavorite: boolean;
   isPremium: boolean;
-  rating: number;
+  location: Location;
   previewImage: string;
+  price: number;
+  title: string;
+  type: string;
 };
